add command line script for calculating multiple force tables at once: useful for sensitivity analysis of w and sigma

+++ src/run_ebmetad/pair_data.py
@@ -26,7 +26,7 @@ class PairData(MetaData):
         probs = self.get('distribution')
         nbins = len(probs)
 
-        force_table = np.zeros(shape=(nbins, nbins))
+        force_table = np.zeros(shape=(nbins, nbins), dtype=np.float32)
 
         # Calculate the effective volume pre-factor
         probs = np.divide(probs, np.sum(probs))  # Normalization, just in case
